Refactor Linux file utilities and update deRegistration logic
