+++ src/utils/linux/fileUtil.js
@@ -1,4 +1,6 @@
 const fs = require('fs');
+const constants = require('../../config/constants');
+const { join } = require('path');
 
 const checkAndRemoveProtocolSchema = (filePaths, protocolSchemaName) => {
     for (const filePath of filePaths) {
@@ -35,8 +37,36 @@ const fileContainsExactLine = (fileContent, searchLine) => {
     return false;
 };
 
+const findRegisteredDesktopFilePath = (defaultApp) => {
+    const desktopFilePaths = [
+        join(constants.osHomeDir, '.local/share/applications', defaultApp)
+    ];
+
+    const xdgEnv = process.env.XDG_DATA_DIRS;
+
+    if (xdgEnv) {
+        const xdgDataDirs = xdgEnv.split(':');
+        if (xdgDataDirs && xdgDataDirs.length) {
+            xdgDataDirs.forEach((path) => {
+                const isApplicationFolderExist = checkIfFolderExists(
+                    join(path, 'applications')
+                );
+
+                if (isApplicationFolderExist) {
+                    desktopFilePaths.push(
+                        join(path, 'applications', defaultApp)
+                    );
+                }
+            });
+        }
+    }
+
+    return desktopFilePaths.find(fs.existsSync);
+};
+
 module.exports = {
     checkAndRemoveProtocolSchema,
     checkIfFolderExists,
-    fileContainsExactLine
+    fileContainsExactLine,
+    findRegisteredDesktopFilePath
 };
